models/disease_identify/requirements.txt
uvicorn==0.22.0
tensorflow==2.19.0
pillow==9.5.0
fastapi==0.115.11
python-multipart==0.0.5
